+++ packages/mpp-solana/CHANGELOG.md
@@ -1,4 +1,4 @@
-# solobank
+# @solobank/mpp-solana
 
 ## 1.0.0